+++ pqcrypto-mlkem/src/mlkem1024.rs
@@ -132,16 +132,13 @@ macro_rules! gen_keypair {
 pub fn keypair() -> (PublicKey, SecretKey) {
     #[cfg(all(enable_x86_avx2, feature = "avx2"))]
     {
-        if std::is_x86_feature_detected!("avx2") {
+        if std::arch::is_x86_feature_detected!("avx2") {
             return gen_keypair!(PQCLEAN_MLKEM1024_AVX2_crypto_kem_keypair);
         }
     }
-    #[cfg(all(enable_aarch64_neon, feature = "neon"))]
+    #[cfg(all(enable_aarch64_sha3, feature = "aarch64-sha3"))]
     {
-        // always use AArch64 code, when target is detected as all AArch64 targets have NEON
-        // support, and std::is_aarch64_feature_detected!("neon") works only with Rust nightly at
-        // the moment
-        if true {
+        if std::arch::is_aarch64_feature_detected!("sha3") {
             return gen_keypair!(PQCLEAN_MLKEM1024_AARCH64_crypto_kem_keypair);
         }
     }
@@ -164,13 +161,13 @@ macro_rules! encap {
 pub fn encapsulate(pk: &PublicKey) -> (SharedSecret, Ciphertext) {
     #[cfg(all(enable_x86_avx2, feature = "avx2"))]
     {
-        if std::is_x86_feature_detected!("avx2") {
+        if std::arch::is_x86_feature_detected!("avx2") {
             return encap!(PQCLEAN_MLKEM1024_AVX2_crypto_kem_enc, pk);
         }
     }
-    #[cfg(all(enable_aarch64_neon, feature = "neon"))]
+    #[cfg(all(enable_aarch64_sha3, feature = "aarch64-sha3"))]
     {
-        if true {
+        if std::arch::is_aarch64_feature_detected!("sha3") {
             return encap!(PQCLEAN_MLKEM1024_AARCH64_crypto_kem_enc, pk);
         }
     }
@@ -192,13 +189,13 @@ macro_rules! decap {
 pub fn decapsulate(ct: &Ciphertext, sk: &SecretKey) -> SharedSecret {
     #[cfg(all(enable_x86_avx2, feature = "avx2"))]
     {
-        if std::is_x86_feature_detected!("avx2") {
+        if std::arch::is_x86_feature_detected!("avx2") {
             return decap!(PQCLEAN_MLKEM1024_AVX2_crypto_kem_dec, ct, sk);
         }
     }
-    #[cfg(all(enable_aarch64_neon, feature = "neon"))]
+    #[cfg(all(enable_aarch64_sha3, feature = "aarch64-sha3"))]
     {
-        if true {
+        if std::arch::is_aarch64_feature_detected!("sha3") {
             return decap!(PQCLEAN_MLKEM1024_AARCH64_crypto_kem_dec, ct, sk);
         }
     }